+++ commands/__init__.py
@@ -1,1 +1,10 @@
-from .reddit_gold import *
+from .donate import donate
+from .hall_of_fame import hall_of_fame
+from .help import help_user
+from .history import history_user
+from .info import info_user
+from .reddit_gold import gold
+from .register import register_user
+from .tip import tip_user
+from .vanity import vanity
+from .withdraw import withdraw_user
